decomposiditon - all in lib
