--- lib/index.js
@@ -1,18 +1,0 @@
-import Vue from 'vue'
-import components from './components'
-// get options passed from module.js
-const options = JSON.parse(`<%= JSON.stringify(options) %>`)
-// loop through components and register them
-for (const name in components) {
-  Vue.component(name, {
-    // extend the original component
-    extends: components[name],
-    // add a _customCounterOptions prop to gain access to the options in the component
-    props: {
-      _customPluginOtions: {
-        type: Object,
-        default: () => ({ ...options })
-      }
-    }
-  })
-}
