+++ lib/module.js
@@ -6,7 +6,7 @@ export default function(moduleOptions) {
   // get all options for the module
   const options = {
     ...moduleOptions,
-    ...this.options.customCounter
+    ...this.options.nuxtPlyr
   }
 
   // expose the namespace / set a default
@@ -15,7 +15,7 @@ export default function(moduleOptions) {
 
   // add all of the initial plugins
   const pluginsToSync = [
-    'lib/index.js',
+    'components/index.js',
   ]
   for (const pathString of pluginsToSync) {
     this.addPlugin({
@@ -26,7 +26,7 @@ export default function(moduleOptions) {
   }
   
   // sync all of the files and folders to revelant places in the nuxt build dir (.nuxt/)
-  const foldersToSync = ['lib/components']
+  const foldersToSync = ['components']
   for (const pathString of foldersToSync) {
     const path = resolve(__dirname, pathString)
     for (const file of readdirSync(path)) {
@@ -38,4 +38,4 @@ export default function(moduleOptions) {
     }
   }
 }
-module.exports.meta = require('./package.json')
+module.exports.meta = require('../package.json')
